503: Added ErrorListItem subcomponent
