+++ src/components/ErrorList/ErrorList.tsx
@@ -1,27 +1,16 @@
 import { useRef, useEffect, useState } from "react";
 import { useTranslation } from "react-i18next";
 
-import { getMessageFromCode } from "eth-rpc-errors";
-
-import {
-  airswapProviderErrorList,
-  ErrorCodesMap,
-  Error,
-  airswapProviderErrorTranslationMap,
-} from "../../constants/errors";
+import type { Error } from "../../constants/errors";
 import useWindowSize from "../../helpers/useWindowSize";
 import { OverlayActionButton } from "../Overlay/Overlay.styles";
-import { InfoHeading } from "../Typography/Typography";
 import {
   Container,
   StyledErrorList,
-  StyledError,
-  ErrorTextContainer,
   LegendDivider,
   StyledScrollContainer,
-  StyledErrorIcon,
-  StyledSubText,
 } from "./ErrorList.styles";
+import ErrorListItem from "./subcomponents/ErrorListItem/ErrorListItem";
 
 type ErrorListProps = {
   errors: Error[];
@@ -35,38 +24,6 @@ export const ErrorList = ({ errors = [], handleClick }: ErrorListProps) => {
   const scrollContainerRef = useRef<HTMLDivElement>(null);
   const [overflow, setOverflow] = useState(false);
   const { width, height } = useWindowSize();
-
-  const StyledErrors = () => {
-    if (!errors.length) return <></>;
-
-    return (
-      <>
-        {errors.map((error, idx) => {
-          const airswapProviderError = airswapProviderErrorList.find(
-            (a) => a === error
-          );
-          const translation = airswapProviderError
-            ? t(
-                `validatorErrors.${airswapProviderErrorTranslationMap[airswapProviderError]}`
-              )
-            : getMessageFromCode(ErrorCodesMap[error], t("common.undefined"));
-
-          return (
-            <StyledError key={idx}>
-              <StyledErrorIcon
-                name="information-circle-outline"
-                iconSize={1.5}
-              />
-              <ErrorTextContainer>
-                <InfoHeading>{error}</InfoHeading>
-                <StyledSubText>{translation}</StyledSubText>
-              </ErrorTextContainer>
-            </StyledError>
-          );
-        })}
-      </>
-    );
-  };
 
   useEffect(() => {
     if (
@@ -88,7 +45,13 @@ export const ErrorList = ({ errors = [], handleClick }: ErrorListProps) => {
       <LegendDivider />
       <StyledScrollContainer $overflow={overflow} ref={scrollContainerRef}>
         <StyledErrorList>
-          <StyledErrors />
+          {
+            errors.map((error) => {
+              return (
+                <ErrorListItem error={error} />
+              )
+            })
+          }
         </StyledErrorList>
       </StyledScrollContainer>
       <OverlayActionButton ref={buttonRef} onClick={handleClick}>
